client/requirements.txt
websockets
SpeechRecognition
PyAudio
pyttsx3
playsound==1.2.2
python-dotenv
openai
deepgram-sdk
elevenlabs
mpv
ffmpeg